+++ src/components/Sidebar/index.ts
@@ -2,5 +2,7 @@ import RvSidebar from "./RvSidebar";
 import RvMenu from "./RvMenu";
 import RvMenuItem from "./RvMenuItem";
 import RvSubMenu from "./RvSubMenu";
+import RvSidebarContainer from "./RvSidebarContainer";
+import RvSidebarHeader from "./RvSidebarHeader";
 
-export { RvSidebar, RvMenu, RvMenuItem, RvSubMenu };
+export { RvSidebar, RvMenu, RvMenuItem, RvSubMenu, RvSidebarContainer, RvSidebarHeader };