+++ src/types/components/Sidebar.ts
@@ -1,5 +1,6 @@
 import type { ReactNode, HTMLAttributes } from "react";
 import type { MenuItemProps, MenuProps, SidebarProps, SubMenuProps } from "react-pro-sidebar";
+import type { StyledDivType, StyledDivHoverType, StyledFlexType } from "./Div";
 
 type BasePropsType = {
 	[props: string]: object | ReactNode | HTMLAttributes<HTMLDivElement>;
@@ -13,3 +14,7 @@ export type RvMenuFCType = MenuProps & BasePropsType;
 export type RvMenuItemFCType = MenuItemProps & BasePropsType;
 
 export type RvSubMenuFCType = SubMenuProps & BasePropsType;
+
+export type RvSidebarContainerFCType = StyledDivType & StyledDivHoverType & BasePropsType;
+
+export type RvSidebarHeaderFCType = StyledFlexType & StyledDivHoverType & BasePropsType;